Update index.html
My idea of the layouts, I went with a drop down on the genres like yours in the beginning but switched to a picture+text kinda thing. I messed around with the date picker a bit but didn't get it to work. But you got it! We can discuss and close/revamp the design. Of course I'd love to beef up the layout even more with some background stuff and etc.